+++ CHANGELOG.rst
@@ -30,3 +30,4 @@ Changelog
 * Added EnergySystem.check() to check graph for sanity
 * Updated build system from setup.py to "build" module
 * Added SubNetwork to build hierachical graphs
+* create_nx_graph(...) now accepts Path objects